# API/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the documentation for the golang api of appolo-api",
        "title": "Appolo API",
        "contact": {
            "name": "API Support",
            "email": "pierre.saintsorny@gmail.com"
        },
        "version": "1"
    },
    "basePath": "/1.0",
    "paths": {
...
        "/account": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get one or multiple account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "my_account",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Account"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "my_account",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "38fzefzef74723jh",
                        "name": "api_key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "38fzefzef74723jh",
                        "name": "secret_key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Delete an account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "my_account",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/binance/account": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get Account Informations about one or multiple asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BTC",
                        "name": "asset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BinanceAccount"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/binance/futures/account": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get Futures Account Informations about one or multiple asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BTC",
                        "name": "asset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "my_account",
                        "name": "account_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FuturesAccount"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
...
        "/data/hourly": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get Hourly data fetch of binance from the scrapper",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BTCUSDT",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "5",
                        "name": "x",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.HourlyDataFetchOnCurrency"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "models.Account": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "insertedDate": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "secretKey": {
                    "type": "string"
                }
            }
        },
        "models.BinanceAccount": {
            "type": "object",
            "properties": {
                "balances": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "asset": {
                                "type": "string"
                            },
                            "free": {
                                "type": "string"
                            },
                            "locked": {
                                "type": "string"
                            }
                        }
                    }
                },
                "canDeposit": {
                    "type": "boolean"
                },
                "canTrade": {
                    "type": "boolean"
                },
                "canWithdraw": {
                    "type": "boolean"
                },
                "makerCommission": {
                    "type": "integer"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "takerCommission": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "integer"
                }
            }
        },
...
        "models.FuturesAccount": {
            "type": "object",
            "properties": {
                "asset": {
                    "type": "string"
                },
                "availableBalance": {
                    "type": "string"
                }
            }
        },
...
        "models.HourlyDataFetchOnCurrency": {
            "type": "object",
            "properties": {
                "bollingerBands": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "chopinessIndex": {
                    "type": "number"
                },
                "close": {
                    "type": "number"
                },
                "closeTime": {
                    "type": "string"
                },
                "currencyName": {
                    "type": "string"
                },
                "high": {
                    "type": "number"
                },
                "ichimoku": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "insertedDate": {
                    "type": "string"
                },
                "low": {
                    "type": "number"
                },
                "ma": {
                    "type": "number"
                },
                "obv": {
                    "type": "number"
                },
                "open": {
                    "type": "number"
                },
                "openTime": {
                    "type": "string"
                },
                "rsi": {
                    "type": "number"
                },
                "sma": {
                    "type": "number"
                },
                "stochastic": {
                    "type": "number"
                },
                "supertrend": {
                    "type": "string"
                },
                "volume": {
                    "type": "number"
                }
            }
        },
...
    }
}